requirements.txt
PyYAML>=6.0
solana>=0.27.0
